refactor(core): extract shared event/comment domain logic into taskify-core package
