+++ taskify-cli/package.json
@@ -36,7 +36,8 @@
     "@nostr-dev-kit/ndk": "^2.18.1",
     "chalk": "^5.3.0",
     "commander": "^12.0.0",
-    "nostr-tools": "^2.16.2"
+    "nostr-tools": "^2.16.2",
+    "taskify-core": "file:../taskify-core"
   },
   "devDependencies": {
     "@types/node": "^25.4.0",
